KS-1-4 new change

+++ src/main/kotlin/lesson2_4/lesson2-task4.kt
@@ -5,8 +5,9 @@ fun main() {
     val age: Byte = 27
     val partOfDay: Float = 0.075f
     val seconds: Short = 6480
-    val partOfYear: Float =  2.0547945205479453E-4f
+    val partOfYear: Double =  2.0547945205479453E-4
     val apogee: Int = 327000
+
     println("Рассотяние - $length")
     println("Возраст - $age")
     println("Часть дня - $partOfDay")
